+++ revision.go
@@ -45,17 +45,20 @@ type RevisionRanges struct {
 
 // RevisionTreeNode is a leaf in Merkle tree with hashed Revisions and with count of documents in the leaf
 type RevisionTreeNode struct {
-	Hash  string `json:"hash"`
-	Count int64  `json:"count,int"`
+	Hash  uint64 `json:"hash"`
+	Count uint64 `json:"count,int"`
 }
 
 // RevisionTree is a list of Revisions in a Merkle tree
 type RevisionTree struct {
-	Version  int                `json:"version"`
-	MaxDepth int                `json:"maxDepth"`
-	RangeMin RevisionUInt64     `json:"rangeMin,string" velocypack:"rangeMin"`
-	RangeMax RevisionUInt64     `json:"rangeMax,string" velocypack:"rangeMax"`
-	Nodes    []RevisionTreeNode `json:"nodes"`
+	Version         int                `json:"version"`
+	MaxDepth        int                `json:"maxDepth"`
+	RangeMin        RevisionUInt64     `json:"rangeMin,string" velocypack:"rangeMin"`
+	RangeMax        RevisionUInt64     `json:"rangeMax,string" velocypack:"rangeMax"`
+	InitialRangeMin RevisionUInt64     `json:"initialRangeMin,string" velocypack:"initialRangeMin"`
+	Count           uint64             `json:"count,int"`
+	Hash            uint64             `json:"hash"`
+	Nodes           []RevisionTreeNode `json:"nodes"`
 }
 
 var (