+++ v2/arangodb/meta.go
@@ -1,7 +1,7 @@
 //
 // DISCLAIMER
 //
-// Copyright 2020 ArangoDB GmbH, Cologne, Germany
+// Copyright 2020-2021 ArangoDB GmbH, Cologne, Germany
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -18,6 +18,7 @@
 // Copyright holder is ArangoDB GmbH, Cologne, Germany
 //
 // Author Adam Janikowski
+// Author Tomasz Mielech
 //
 
 package arangodb
@@ -26,6 +27,12 @@ import (
 	"github.com/arangodb/go-driver/v2/arangodb/shared"
 	"github.com/pkg/errors"
 )
+
+// ShardID is an internal identifier of a specific shard.
+type ShardID string
+
+// ServerID identifies an ArangoDB server in a cluster.
+type ServerID string
 
 type DocumentID string