+++ v2/connection/connection_http_bytes.go
@@ -1,7 +1,7 @@
 //
 // DISCLAIMER
 //
-// Copyright 2020-2021 ArangoDB GmbH, Cologne, Germany
+// Copyright 2021 ArangoDB GmbH, Cologne, Germany
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -17,24 +17,13 @@
 //
 // Copyright holder is ArangoDB GmbH, Cologne, Germany
 //
-// Author Adam Janikowski
 // Author Tomasz Mielech
 //
 
-package arangodb
+package connection
 
-import "context"
-
-type Collection interface {
-	Name() string
-	Database() Database
-
-	// Shards fetches shards information of the collection.
-	Shards(ctx context.Context, details bool) (CollectionShards, error)
-
-	// Remove removes the entire collection.
-	// If the collection does not exist, a NotFoundError is returned.
-	Remove(ctx context.Context) error
-
-	CollectionDocuments
-}
+const (
+	PlainText              = "text/plain"
+	ApplicationOctetStream = "application/octet-stream"
+	ApplicationZip         = "application/zip"
+)