+++ v2/connection/connection_http_internal.go
@@ -1,7 +1,7 @@
 //
 // DISCLAIMER
 //
-// Copyright 2020 ArangoDB GmbH, Cologne, Germany
+// Copyright 2020-2021 ArangoDB GmbH, Cologne, Germany
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -18,6 +18,7 @@
 // Copyright holder is ArangoDB GmbH, Cologne, Germany
 //
 // Author Adam Janikowski
+// Author Tomasz Mielech
 //
 
 package connection
@@ -102,30 +103,21 @@ func (j *httpConnection) SetAuthentication(a Authentication) error {
 	return nil
 }
 
-func (j httpConnection) Decoder(content string) Decoder {
-	switch content {
-	case ApplicationVPack:
-		return getVPackDecoder()
-	case ApplicationJSON:
-		return getJsonDecoder()
-	default:
-		switch j.contentType {
-		case ApplicationVPack:
-			return getVPackDecoder()
-		case ApplicationJSON:
-			return getJsonDecoder()
-		default:
-			return getJsonDecoder()
-		}
+// Decoder returns the decoder according to the response content type or HTTP connection request content type.
+// If the content type is unknown then it returns default JSON decoder.
+func (j httpConnection) Decoder(contentType string) Decoder {
+	// First try to get decoder by the content type of the response.
+	if decoder := getDecoderByContentType(contentType); decoder != nil {
+		return decoder
 	}
-}
 
-func (j httpConnection) DoWithReader(ctx context.Context, request Request) (Response, io.ReadCloser, error) {
-	req, ok := request.(*httpRequest)
-	if !ok {
-		return nil, nil, errors.Errorf("unable to parse request into JSON Request")
+	// Next try to get decoder by the content type of the HTTP connection.
+	if decoder := getDecoderByContentType(j.contentType); decoder != nil {
+		return decoder
 	}
-	return j.do(ctx, req)
+
+	// Return the default decoder.
+	return getJsonDecoder()
 }
 
 func (j httpConnection) GetEndpoint() Endpoint {
@@ -170,37 +162,44 @@ func (j httpConnection) newRequestWithEndpoint(endpoint string, method string, u
 	return r, nil
 }
 
+// Do performs HTTP request and returns the response.
+// If `output` is provided then it is populated from response body and the response is automatically freed.
 func (j httpConnection) Do(ctx context.Context, request Request, output interface{}) (Response, error) {
-	req, ok := request.(*httpRequest)
-	if !ok {
-		return nil, errors.Errorf("unable to parse request into JSON Request")
-	}
-	return j.doWithOutput(ctx, req, output)
-}
-
-func (j httpConnection) doWithOutput(ctx context.Context, request *httpRequest, output interface{}) (*httpResponse, error) {
-	resp, body, err := j.do(ctx, request)
+	resp, body, err := j.Stream(ctx, request)
 	if err != nil {
 		return nil, err
 	}
 
-	if output != nil {
-		defer dropBodyData(body) // In case if there is data drop it all
+	// The body should be closed at the end of the function.
+	defer dropBodyData(body)
 
+	if output != nil {
+		// The output should be stored in the output variable.
 		if err = j.Decoder(resp.Content()).Decode(body, output); err != nil {
 			if err != io.EOF {
 				return nil, errors.WithStack(err)
 			}
 		}
-	} else {
-		// We still need to read data from request, but we can do this in background and ignore output
-		defer dropBodyData(body)
 	}
 
 	return resp, nil
 }
 
-func (j httpConnection) do(ctx context.Context, req *httpRequest) (*httpResponse, io.ReadCloser, error) {
+// Stream performs HTTP request.
+// It returns the response and body reader to read the data from there.
+// The caller is responsible to free the response body.
+func (j httpConnection) Stream(ctx context.Context, request Request) (Response, io.ReadCloser, error) {
+	req, ok := request.(*httpRequest)
+	if !ok {
+		return nil, nil, errors.Errorf("unable to parse request into JSON Request")
+	}
+
+	return j.stream(ctx, req)
+}
+
+// stream performs the HTTP request.
+// It returns HTTP response and body reader to read the data from there.
+func (j httpConnection) stream(ctx context.Context, req *httpRequest) (*httpResponse, io.ReadCloser, error) {
 	id := uuid.New().String()
 	log.Debugf("(%s) Sending request to %s/%s", id, req.Method(), req.URL())
 	if v, ok := req.GetHeader(ContentType); !ok || v == "" {
@@ -222,6 +221,7 @@ func (j httpConnection) do(ctx context.Context, req *httpRequest) (*httpResponse
 		ctx = context.Background()
 	}
 
+	var bodyReader io.Reader
 	if req.Method() == http.MethodPost || req.Method() == http.MethodPut || req.Method() == http.MethodPatch {
 		decoder := j.Decoder(j.contentType)
 		if !j.streamSender {
@@ -230,12 +230,7 @@ func (j httpConnection) do(ctx context.Context, req *httpRequest) (*httpResponse
 				return nil, nil, err
 			}
 
-			r, err := req.asRequest(ctx, b)
-			if err != nil {
-				return nil, nil, errors.WithStack(err)
-			}
-
-			httpReq = r
+			bodyReader = b
 		} else {
 			reader, writer := io.Pipe()
 			go func() {
@@ -245,20 +240,15 @@ func (j httpConnection) do(ctx context.Context, req *httpRequest) (*httpResponse
 				}
 			}()
 
-			r, err := req.asRequest(ctx, reader)
-			if err != nil {
-				return nil, nil, errors.WithStack(err)
-			}
-
-			httpReq = r
+			bodyReader = reader
 		}
-	} else {
-		r, err := req.asRequest(ctx, nil)
-		if err != nil {
-			return nil, nil, errors.WithStack(err)
-		}
-		httpReq = r
 	}
+
+	r, err := req.asRequest(ctx, bodyReader)
+	if err != nil {
+		return nil, nil, errors.WithStack(err)
+	}
+	httpReq = r
 
 	resp, err := j.client.Do(httpReq)
 	if err != nil {
@@ -275,4 +265,19 @@ func (j httpConnection) do(ctx context.Context, req *httpRequest) (*httpResponse
 	}
 
 	return &httpResponse{response: resp, request: req}, nil, nil
+}
+
+// getDecoderByContentType returns the decoder according to the content type.
+// If content type is unknown then nil is returned.
+func getDecoderByContentType(contentType string) Decoder {
+	switch contentType {
+	case ApplicationVPack:
+		return getVPackDecoder()
+	case ApplicationJSON:
+		return getJsonDecoder()
+	case PlainText, ApplicationOctetStream, ApplicationZip:
+		return getBytesDecoder()
+	default:
+		return nil
+	}
 }