+++ CHANGELOG.md
@@ -1,10 +1,21 @@
 # Change Log
 
 ## [master](https://github.com/arangodb/go-driver/tree/master) (N/A)
+- [V2] Add support for streaming the response body by the caller.
+
+## [1.2.1](https://github.com/arangodb/go-driver/tree/v1.2.1) (2021-09-21)
+- Add support for fetching shards' info by the given collection name.
+- Change versioning to be go mod compatible
+- Add support for ForceOneShardAttributeValue in Query
+
+## [1.2.0](https://github.com/arangodb/go-driver/tree/1.2.0) (2021-08-04)
 - Add support for AQL, Pipeline, Stopwords, GeoJSON and GeoPoint Arango Search analyzers.
-- Add `estimates` field to indexes properties
-- Add tests for 3.8 ArangoDB and remove tests for 3.5
-- Add Plan support in Query execution
+- Add `estimates` field to indexes properties.
+- Add tests for 3.8 ArangoDB and remove tests for 3.5.
+- Add Plan support in Query execution.
+- Change Golang version from 1.13.4 to 1.16.6.
+- Add graceful shutdown for the coordinators.
+- Replace 'github.com/dgrijalva/jwt-go' with 'github.com/golang-jwt/jwt'
 
 ## [1.1.1](https://github.com/arangodb/go-driver/tree/1.1.1) (2020-11-13)
 - Add Driver V2 in Alpha version